patch: refactor webracker company enrichment
