+++ .gitignore
@@ -55,3 +55,4 @@ priv/scripts/*
 core.iml
 *.mdc
 .cursorrules
+CLAUDE.md
